Fix the pre-check quota issue for rate sensor.
For rate sensor the value we used to compare with quota should be usage/elapsed. So once we do the pre-check, we should aslo use the same rule to compute the request value.No the value we used to compare with quota should be (usage+requestValue)/elapsed.

Add javadoc and refactor method name based on review comment.