Try to avoid `synchronized` in `Sensor#record`, which is a hot path
Originally, the `synchronized` section was introduced by quota logic,
and it is in the critical path for every record invocation.
Considering quota config is only required in every few sensors, it
doesn't make sense to slow down the hot path, and this code change
adds a check before entering the `synchronized` section to remove
the contention for regular metrics.