AsyncGauge returns error code for when measurement breaches max timeout
By default, the error code is -1, but it will be configurable.

Other changes:
Instead of blocking in the record() function in AsyncGauge, change it
to do nothing, because this stat can be used together with other stats
which support record(), and record() will be called for all the stats
in the same Sensor.