Added DaemonThreadFactory to avoid hanging in the tehuti application
Previously, 'RedundantLogFilter' is using non-daemon thread to clean up
the log filter, and it would result in stuck application if the application
doesn't invoke 'RedundantLogFilter#shutdown', which is error prone.
This PR adds a 'DaemonThreadFactory', so that the log filter threads
won't block Tehuti application shutdown even it doesn't invoke
'RedundantLogFilter#shutdown' explicitly.