Fixed the order of locking between MetricRepository and Sensor to prevent deadlock (#34)
Currently lock on MetricRepository is taken first and then Sensor's lock is taken in some cases and the opposite is happening is some other cases, leading to deadlock if both these cases happen around the same time. Fix is to change the order of locking to always take lock on MetricRepository first and then take Sensor's lock whenever the MetricRepository is accessed in Sensor to prevent deadlock.